+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/update/flush/ParentReferenceAttributeFlusher.java
@@ -95,7 +95,7 @@ public class ParentReferenceAttributeFlusher<E, V> extends BasicAttributeFlusher
     }
 
     @Override
-    public Query flushQuery(UpdateContext context, String parameterPrefix, UpdateQueryFactory queryFactory, Query query, Object ownerView, Object view, V value, UnmappedOwnerAwareDeleter ownerAwareDeleter) {
+    public Query flushQuery(UpdateContext context, String parameterPrefix, UpdateQueryFactory queryFactory, Query query, Object ownerView, Object view, V value, UnmappedOwnerAwareDeleter ownerAwareDeleter, DirtyAttributeFlusher<?, ?, ?> ownerFlusher) {
         if (query != null && writableMappings != null) {
             for (int i = 0; i < parameterAccessors.length; i++) {
                 Map.Entry<String, AttributeAccessor> parameterAccessor = parameterAccessors[i];
@@ -108,7 +108,7 @@ public class ParentReferenceAttributeFlusher<E, V> extends BasicAttributeFlusher
                 query.setParameter(parameter, parameterAccessor.getValue().getValue(value));
             }
         } else {
-            super.flushQuery(context, parameterPrefix, queryFactory, query, ownerView, view, value, ownerAwareDeleter);
+            super.flushQuery(context, parameterPrefix, queryFactory, query, ownerView, view, value, ownerAwareDeleter, ownerFlusher);
         }
         return query;
     }