+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/update/flush/PersistCollectionElementAttributeFlusher.java
@@ -48,7 +48,7 @@ public class PersistCollectionElementAttributeFlusher<E, V> extends CollectionEl
     }
 
     @Override
-    public Query flushQuery(UpdateContext context, String parameterPrefix, UpdateQueryFactory queryFactory, Query query, Object ownerView, Object view, V value, UnmappedOwnerAwareDeleter ownerAwareDeleter) {
+    public Query flushQuery(UpdateContext context, String parameterPrefix, UpdateQueryFactory queryFactory, Query query, Object ownerView, Object view, V value, UnmappedOwnerAwareDeleter ownerAwareDeleter, DirtyAttributeFlusher<?, ?, ?> ownerFlusher) {
         context.getEntityManager().persist(element);
         return query;
     }
@@ -61,7 +61,7 @@ public class PersistCollectionElementAttributeFlusher<E, V> extends CollectionEl
     }
 
     @Override
-    public DirtyAttributeFlusher<CollectionElementAttributeFlusher<E, V>, E, V> getDirtyFlusher(UpdateContext context, Object view, Object initial, Object current, List<Runnable> preFlushListeners) {
+    public DirtyAttributeFlusher<CollectionElementAttributeFlusher<E, V>, E, V> getDirtyFlusher(UpdateContext context, Object view, Object initial, Object current) {
         // Actually this should never be called, but let's return this to be safe
         return this;
     }