+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/update/flush/SubviewAttributeFlusher.java
@@ -185,7 +185,7 @@ public class SubviewAttributeFlusher<E, V> extends AttributeFetchGraphNode<Subvi
     }
 
     @Override
-    public Query flushQuery(UpdateContext context, String parameterPrefix, UpdateQueryFactory queryFactory, Query query, Object ownerView, Object view, V value, UnmappedOwnerAwareDeleter ownerAwareDeleter) {
+    public Query flushQuery(UpdateContext context, String parameterPrefix, UpdateQueryFactory queryFactory, Query query, Object ownerView, Object view, V value, UnmappedOwnerAwareDeleter ownerAwareDeleter, DirtyAttributeFlusher<?, ?, ?> ownerFlusher) {
         V finalValue;
         if (flushOperation == null) {
             finalValue = value;
@@ -219,18 +219,12 @@ public class SubviewAttributeFlusher<E, V> extends AttributeFetchGraphNode<Subvi
                     }
                 } else {
                     int orphanRemovalStartIndex = context.getOrphanRemovalDeleters().size();
-                    try {
-                        Query q = nestedFlusher.flushQuery(context, parameterPrefix, viewToEntityMapper, null, finalValue, null, finalValue, ownerAwareDeleter);
-                        if (q != null) {
-                            int updated = q.executeUpdate();
+                    Query q = nestedFlusher.flushQuery(context, parameterPrefix, viewToEntityMapper, null, finalValue, null, finalValue, ownerAwareDeleter, nestedFlusher);
+                    if (q != null) {
+                        int updated = q.executeUpdate();
 
-                            if (updated != 1) {
-                                throw new OptimisticLockException(null, finalValue);
-                            }
-                        }
-                    } finally {
-                        if (finalValue instanceof MutableStateTrackable) {
-                            context.getInitialStateResetter().addUpdatedView((MutableStateTrackable) finalValue);
+                        if (updated != 1) {
+                            throw new OptimisticLockException(null, finalValue);
                         }
                     }
                     context.removeOrphans(orphanRemovalStartIndex);
@@ -245,9 +239,9 @@ public class SubviewAttributeFlusher<E, V> extends AttributeFetchGraphNode<Subvi
                     Object realValue = v == null ? null : subviewIdAccessor.getValue(finalValue);
                     if (supportElementIdQueryFlush) {
                         if (parameterPrefix == null) {
-                            getElementIdFlusher().flushQuery(context, parameterName + "_", queryFactory, query, finalValue, null, realValue, ownerAwareDeleter);
+                            getElementIdFlusher().flushQuery(context, parameterName + "_", queryFactory, query, finalValue, null, realValue, ownerAwareDeleter, null);
                         } else {
-                            getElementIdFlusher().flushQuery(context, parameterPrefix + parameterName + "_", queryFactory, query, finalValue, null, realValue, ownerAwareDeleter);
+                            getElementIdFlusher().flushQuery(context, parameterPrefix + parameterName + "_", queryFactory, query, finalValue, null, realValue, ownerAwareDeleter, null);
                         }
                     } else {
                         if (parameterPrefix == null) {
@@ -274,18 +268,12 @@ public class SubviewAttributeFlusher<E, V> extends AttributeFetchGraphNode<Subvi
                     }
                 } else {
                     int orphanRemovalStartIndex = context.getOrphanRemovalDeleters().size();
-                    try {
-                        Query q = nestedFlusher.flushQuery(context, parameterPrefix, viewToEntityMapper, null, value, null, value, ownerAwareDeleter);
-                        if (q != null) {
-                            int updated = q.executeUpdate();
+                    Query q = nestedFlusher.flushQuery(context, parameterPrefix, viewToEntityMapper, null, value, null, value, ownerAwareDeleter, nestedFlusher);
+                    if (q != null) {
+                        int updated = q.executeUpdate();
 
-                            if (updated != 1) {
-                                throw new OptimisticLockException(null, value);
-                            }
-                        }
-                    } finally {
-                        if (value instanceof MutableStateTrackable) {
-                            context.getInitialStateResetter().addUpdatedView((MutableStateTrackable) value);
+                        if (updated != 1) {
+                            throw new OptimisticLockException(null, value);
                         }
                     }
                     context.removeOrphans(orphanRemovalStartIndex);
@@ -299,9 +287,9 @@ public class SubviewAttributeFlusher<E, V> extends AttributeFetchGraphNode<Subvi
                     Object realValue = v == null ? null : subviewIdAccessor.getValue(value);
                     if (supportElementIdQueryFlush) {
                         if (parameterPrefix == null) {
-                            getElementIdFlusher().flushQuery(context, parameterName + "_", queryFactory, query, value, null, realValue, ownerAwareDeleter);
+                            getElementIdFlusher().flushQuery(context, parameterName + "_", queryFactory, query, value, null, realValue, ownerAwareDeleter, null);
                         } else {
-                            getElementIdFlusher().flushQuery(context, parameterPrefix + parameterName + "_", queryFactory, query, value, null, realValue, ownerAwareDeleter);
+                            getElementIdFlusher().flushQuery(context, parameterPrefix + parameterName + "_", queryFactory, query, value, null, realValue, ownerAwareDeleter, null);
                         }
                     } else {
                         if (parameterPrefix == null) {
@@ -326,18 +314,12 @@ public class SubviewAttributeFlusher<E, V> extends AttributeFetchGraphNode<Subvi
             }
             if (nestedFlusher != null && nestedFlusher != viewToEntityMapper.getFullGraphNode()) {
                 int orphanRemovalStartIndex = context.getOrphanRemovalDeleters().size();
-                try {
-                    Query q = nestedFlusher.flushQuery(context, parameterPrefix, viewToEntityMapper, null, realValue, null, realValue, ownerAwareDeleter);
-                    if (q != null) {
-                        int updated = q.executeUpdate();
+                Query q = nestedFlusher.flushQuery(context, parameterPrefix, viewToEntityMapper, null, realValue, null, realValue, ownerAwareDeleter, nestedFlusher);
+                if (q != null) {
+                    int updated = q.executeUpdate();
 
-                        if (updated != 1) {
-                            throw new OptimisticLockException(null, realValue);
-                        }
-                    }
-                } finally {
-                    if (realValue instanceof MutableStateTrackable) {
-                        context.getInitialStateResetter().addUpdatedView((MutableStateTrackable) realValue);
+                    if (updated != 1) {
+                        throw new OptimisticLockException(null, realValue);
                     }
                 }
                 context.removeOrphans(orphanRemovalStartIndex);
@@ -387,13 +369,7 @@ public class SubviewAttributeFlusher<E, V> extends AttributeFetchGraphNode<Subvi
                 if (update && inverseFlusher != null) {
                     inverseFlusher.flushEntitySetElement(context, finalValue, entity, entity, (DirtyAttributeFlusher<?, E, Object>) (DirtyAttributeFlusher<?, ?, ?>) nestedFlusher);
                 } else {
-                    try {
-                        nestedFlusher.flushEntity(context, null, finalValue, null, finalValue, null);
-                    } finally {
-                        if (finalValue instanceof MutableStateTrackable) {
-                            context.getInitialStateResetter().addUpdatedView((MutableStateTrackable) finalValue);
-                        }
-                    }
+                    nestedFlusher.flushEntity(context, null, finalValue, null, finalValue, null);
                 }
             } else if (update && inverseFlusher != null) {
                 if (finalValue instanceof MutableStateTrackable) {
@@ -421,13 +397,7 @@ public class SubviewAttributeFlusher<E, V> extends AttributeFetchGraphNode<Subvi
                     inverseFlusher.flushEntitySetElement(context, finalValue, entity, entity, (DirtyAttributeFlusher<?, E, Object>) (DirtyAttributeFlusher<?, ?, ?>) nestedFlusher);
                     wasDirty |= true;
                 } else {
-                    try {
-                        wasDirty |= nestedFlusher.flushEntity(context, null, finalValue, null, finalValue, null);
-                    } finally {
-                        if (finalValue instanceof MutableStateTrackable) {
-                            context.getInitialStateResetter().addUpdatedView((MutableStateTrackable) finalValue);
-                        }
-                    }
+                    wasDirty |= nestedFlusher.flushEntity(context, null, finalValue, null, finalValue, null);
                 }
             } else if (update && inverseFlusher != null) {
                 inverseFlusher.flushEntitySetElement(context, finalValue, entity, entity, null);
@@ -465,13 +435,7 @@ public class SubviewAttributeFlusher<E, V> extends AttributeFetchGraphNode<Subvi
         } else {
             V realValue = (V) viewAttributeAccessor.getValue(view);
             if (nestedFlusher != null && nestedFlusher != viewToEntityMapper.getFullGraphNode()) {
-                try {
-                    wasDirty |= nestedFlusher.flushEntity(context, null, realValue, null, realValue, null);
-                } finally {
-                    if (realValue instanceof MutableStateTrackable) {
-                        context.getInitialStateResetter().addUpdatedView((MutableStateTrackable) realValue);
-                    }
-                }
+                wasDirty |= nestedFlusher.flushEntity(context, null, realValue, null, realValue, null);
             } else {
                 // We don't care if the underlying real value is different from the initial value if we don't have an entity attribute accessor
                 // That represents a correlated attribute so we will only do cascading
@@ -626,7 +590,7 @@ public class SubviewAttributeFlusher<E, V> extends AttributeFetchGraphNode<Subvi
 
     @Override
     @SuppressWarnings("unchecked")
-    public DirtyAttributeFlusher<SubviewAttributeFlusher<E, V>, E, V> getDirtyFlusher(UpdateContext context, Object view, Object initial, Object current, List<Runnable> preFlushListeners) {
+    public DirtyAttributeFlusher<SubviewAttributeFlusher<E, V>, E, V> getDirtyFlusher(UpdateContext context, Object view, Object initial, Object current) {
         if (isPassThrough()) {
             return null;
         }